Add two small fixes to verticaltabs.css
If the tabs are vertical, the pinned tabs don't need to be small,
because the whole row is reserved for them anyways, so we unset the
max-width and width CSS properties.

But this results in a new UI fail: the pinned tabs are not easy to
say apart from non-pinned tabs, so to battle this, we add a slightly
darker color as their background (but only if they are not actually
currently selected, in which case we keep the color of the "cursor").